add specific implementation for netstd2.1
