+++ .editorconfig
@@ -1,0 +1,3 @@
+[*.cs]
+indent_size = 4
+indent_style = space
