+++ PooledStream/PooledStream.csproj
@@ -1,7 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard1.1</TargetFramework>
+    <TargetFrameworks>netstandard1.1;netstandard2.1</TargetFrameworks>
+    <!-- <TargetFramework>netstandard2.1</TargetFramework> -->
     <PackageId>PooledStream</PackageId>
     <PackageVersion>0.1.0</PackageVersion>
     <Authors>itn3000</Authors>
@@ -16,7 +17,7 @@
   </PropertyGroup>
   <ItemGroup>
   </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="System.Buffers" Version="4.3.0" />
+  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard1.1'">
+    <PackageReference Include="System.Buffers" Version="4.5.0" />
   </ItemGroup>
 </Project>
